feat(board): display current hand using a dialog
