+++ src/app/components/card/services/card.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable, inject } from '@angular/core';
 import { withCache } from '@ngneat/cashew';
-import { Observable, catchError, of, shareReplay } from 'rxjs';
+import { Observable, catchError, of } from 'rxjs';
 import { environment } from 'src/environments';
 import { CardDto } from '../dtos';
 
@@ -10,10 +10,6 @@ export class CardService {
   readonly #httpClient = inject(HttpClient);
 
   getCardById(id: string): Observable<CardDto> {
-    // TODO: Remove when there is a real API
-    // eslint-disable-next-line no-console
-    console.log('Getting card by ID:', id);
-
     return (
       this.#httpClient
         // TODO: Validate URL when there is a real API
@@ -29,7 +25,6 @@ export class CardService {
 
             return of({ id, url: 'assets/images/back-card.webp' });
           }),
-          shareReplay({ bufferSize: 1, refCount: true }),
         )
     );
   }
