+++ src/app/components/game-zones/hand-zone/hand-opened/index.ts
