+++ src/app/components/game-zones/hand-zone/services/index.ts
@@ -1,1 +1,2 @@
+export * from './current-hand.service';
 export * from './display-hand.service';
