--- src/app/components/game-zones/hand/hand-opened/hand-opened.component.html
@@ -1,9 +1,0 @@
-<div
-  class="no-scrollbar flex h-[110%] items-center gap-x-3 overflow-x-scroll px-8"
->
-  @for (cardId of cardIds; track cardId) {
-    <div class="hover:scale-110">
-      <app-card [cardId]="cardId" size="lg" />
-    </div>
-  }
-</div>
